refactor(calendar): lift chunkArray into core/utils
Moves the Google-local chunkArray helper into core/utils/chunk so CalDAV can
reuse it without importing from providers/google. No behaviour change.
