+++ packages/calendar/src/core/utils/chunk.ts
@@ -1,0 +1,9 @@
+const chunkArray = <TItem>(items: TItem[], size: number): TItem[][] => {
+  const chunks: TItem[][] = [];
+  for (let offset = 0; offset < items.length; offset += size) {
+    chunks.push(items.slice(offset, offset + size));
+  }
+  return chunks;
+};
+
+export { chunkArray };
